Rewriting the drainaquifer script
The original drainaquifer script could occasionally fail to notice aquifer layers if they happen to skip each of the special tiles that it checks, two per block.  It also left the block-level aquifer flags set, which seems like a minor waste of FPS.

The new script uses the block-level flags to determine which blocks to check, which seems to be both more correct and faster.  It's also written in Lua instead of Ruby, for clarity and ease of debugging.  The name has changed slightly to prevent problems if both scripts are installed for some reason.