Autogen script links, table of contents, and demo!
Automatically generating the link target for each script saves a lot of
time and potential for errors.

Each kind-of-script page now includes a link target for that page, and
also a table of contents.

In-script markers to delimit text for Sphinx extraction now use native
docstring markers from Ruby, which is a lot more elegant in .rb scripts
- and robust, as long as the fisrt docstring is valid .rst!