fixed a few issues
occasionally the original point of creation has no map data (its below the bottom floor of hell, and therefore does not exist).
figured that since the point it tries to create the unit is a series of simulated actions that moves through the arena spawn unit menus, that before opening the menu, you could just move the cursor there, if location was set, it worked.  

also added a argument to set the unit to civ_id and group_id of the the fort, to reduce on the verbose of most commands (nearly every use I have for create-unit is set to the civ and group of the fort...).  

moved the civ_id change closer to the point of actually creating the unit,  thought about moving it into the createNemesis function... but didn't.

corrected error involving out of bounds vector, on old forts in the setting of the age.  It was just a bad way to look up the historical figure, used the appropriate call.