Added bodyswap.lua
A revived version of the long defunct "adv-bodyswap" with added functionality. 
This script allows the player to swap their control over to any unit present on the map in adventure mode only. Targets are specified either via their unit id or through UI selection prior to running the script. Unlike previous iterations, units lacking nemesis or historical figure data (such as wild animals) are valid and safe targets, as the script creates this data when it is missing.