Fix & Update for v0.44.12
Old version would cause a consistent DF crash whenever it was used, and then a particular dwarf's thoughts and preferences were examined.  This was largely due to not handling music, dances, and poetic values properly.

This version fixes that, and adds several new features.  Notable is the ability to put together preference profiles pretty easily, as demonstrated by the 'goth' option.  If there's enough interest, I'll add a preference profile generator with persistent storage.
I also updated the documentation block so that it reflects reality, now.  The 'list' option is very handy, and enumerates all the songs, dances, and poems in the world, among other preference types.