Update leave to properly remove companion links
This change puts leave closer to the vanilla game logic, since companion links are also severed when dismissing through normal means ingame. Previously, any companion dismissed via leave would still be listed as a travelling companion in the adventurer's description screen because the link wasn't removed. This update still doesn't cancel the adventuring agreement, and so it will still be listed as active in the log, though that doesn't prevent the player for re-recruiting that companion. I need to figure out how to set an agreement as ended...