implement creation of any kind of manager_order
The core of this script now can process a manager order defined in a Lua-table and create an actual order from it. It is translated 1-to-1 from orders.cpp to Lua (modified very slightly to use this as planning, not "hacking" tool). Additionally, it is possible to use JSON strings or files.

Few advanced usage examples are in the help; also use `orders export filename` for inspiration.