rewrite extent splitting code
I found a pathological case during playtesting. when a flood fill
discovers connected buildings of the same type, but the area is greater
than the max dimensions of the building type *and* there is a hole in
the extent, the previous code erroneously filled in the hole. This
caused, for example, a "C"-shaped string of beds to become a solid block
of beds.

Another difficult case was when a group of 3x3 workshops were arranged
in a solid block, but the workshops were not aligned in a grid. the
extent splitting code forcibly rearranged the workshops into an aligned
grid : )

Both cases work now, and I've refactored the vertical and horizontal
extent splitting code so future fixes only have to be made in one place.