Switch to unit naming mode
Necessary following the recent switch to viewscreen_setupdwarfgamest. Without the naming type set correctly, hitting "randomise" will generate site names rather than unit names, and the "first name" option is greyed out. The magic number should be switched with a proper enum when this becomes available. If a different mode for generating item names is found, this should be set when naming artifacts.