always create priority event block
this prevents the case where a blueprint has some priorities specified
but not others, and the default priorities are scanned first. the event
block would not get created for the default priorities, then when the
ones with priority set are scanned, the event block would get created
and initialized to 0. the default priority tiles would then have a
priority of 0, which is invalid.